JP-2707: Extract comserver components from com.energyict.protocol.messaging into protocols bundle

+++ comu/com.energyict.mdc.protocol.pluggable/src/main/java/com/energyict/mdc/protocol/pluggable/impl/adapters/smartmeterprotocol/SmartMeterProtocolLoadProfileAdapter.java
@@ -171,7 +171,7 @@ public class SmartMeterProtocolLoadProfileAdapter implements DeviceLoadProfileSu
      * @param loadProfileId   the ID to look for
      * @return the requested ProfileData or the {@link #INVALID_PROFILE_DATA} when the loadProfileId is not found
      */
-    protected ProfileData getProfileDataWithLoadProfileId(final List<ProfileData> profileDataList, int loadProfileId) {
+    protected ProfileData getProfileDataWithLoadProfileId(final List<ProfileData> profileDataList, long loadProfileId) {
         if (profileDataList != null) {
             for (ProfileData profileData : profileDataList) {
                 if (profileData.getLoadProfileId() == loadProfileId) {
